## Description
This speeds up the slowest tests, AutoType and Kdbx4, by adjusting some parameters which should not influence the tested semantics.

## Motivation and context
Tests are slow. It's more fun to write code with faster test suites.

## How has this been tested?
I ran tests on Debian GNU/Linux on an i7-6700K with ASAN on and off. The times went down from approximately 9.71s and 8.72s respectively to 4.38s and 3.08s for a `make test`. `TestKdbx4` is still slow (still lots of cryptography), but `TestAutoType` is almost negligible.

## Types of changes
- Speed optimization

## Checklist:
- ✅ I have read the **CONTRIBUTING** document. **[REQUIRED]**
- ✅ My code follows the code style of this project. **[REQUIRED]**
- ✅ All new and existing tests passed. **[REQUIRED]**
- ✅ I have compiled and verified my code with `-DWITH_ASAN=ON`. **[REQUIRED]**